chore: reorg codes to helper
